+++ lib/src/main/java/com/diffplug/spotless/npm/PrettierFormatterStep.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016-2021 DiffPlug
+ * Copyright 2016-2022 DiffPlug
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -23,10 +23,11 @@ import java.io.Serializable;
 import java.util.Collections;
 import java.util.Map;
 import java.util.TreeMap;
-import java.util.logging.Level;
-import java.util.logging.Logger;
 
 import javax.annotation.Nonnull;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import com.diffplug.spotless.FormatterFunc;
 import com.diffplug.spotless.FormatterFunc.Closeable;
@@ -36,7 +37,7 @@ import com.diffplug.spotless.ThrowingEx;
 
 public class PrettierFormatterStep {
 
-	private static final Logger logger = Logger.getLogger(PrettierFormatterStep.class.getName());
+	private static final Logger logger = LoggerFactory.getLogger(PrettierFormatterStep.class);
 
 	public static final String NAME = "prettier-format";
 
@@ -95,7 +96,7 @@ public class PrettierFormatterStep {
 			try {
 				restService.shutdown();
 			} catch (Throwable t) {
-				logger.log(Level.INFO, "Failed to request shutdown of rest service via api. Trying via process.", t);
+				logger.info("Failed to request shutdown of rest service via api. Trying via process.", t);
 			}
 			restServer.close();
 		}